+++ apps/dashboard/src/contract-ui/tabs/account/components/deposit-native.tsx
@@ -3,27 +3,28 @@ import { TransactionButton } from "components/buttons/TransactionButton";
 import { thirdwebClient } from "lib/thirdweb-client";
 import { type ChangeEvent, useState } from "react";
 import { prepareTransaction, toWei } from "thirdweb";
-import {
-  useActiveWalletChain,
-  useSendAndConfirmTransaction,
-} from "thirdweb/react";
+import { useSendAndConfirmTransaction } from "thirdweb/react";
 import { Card } from "tw-components";
+import type { StoredChain } from "../../../../contexts/configured-chains";
+import { mapStoredChainTov5Chain } from "../../../../contexts/map-chains";
 
 interface DepositNativeProps {
   address: string;
   symbol: string;
+  chain: StoredChain;
 }
 
 export const DepositNative: React.FC<DepositNativeProps> = ({
   address,
   symbol,
+  chain,
 }) => {
-  const chain = useActiveWalletChain();
   const { mutate: transfer, isPending } = useSendAndConfirmTransaction();
   const [amount, setAmount] = useState("");
   const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
     setAmount(e.currentTarget.value);
   };
+
   return (
     <Card
       style={{
@@ -44,22 +45,17 @@ export const DepositNative: React.FC<DepositNativeProps> = ({
         transactionCount={1}
         isLoading={isPending}
         isDisabled={
-          amount.length === 0 ||
-          Number.parseFloat(amount) <= 0 ||
-          !address ||
-          !chain
+          amount.length === 0 || Number.parseFloat(amount) <= 0 || !address
         }
         colorScheme="primary"
         onClick={() => {
           if (!address) {
             throw new Error("Invalid address");
           }
-          if (!chain) {
-            throw new Error("Invalid chain");
-          }
+
           const transaction = prepareTransaction({
             to: address,
-            chain,
+            chain: mapStoredChainTov5Chain(chain),
             client: thirdwebClient,
             value: toWei(amount),
           });
